refactor: improve code readability
