+++ src/server/routes/unifiedform.ts
@@ -1,8 +1,8 @@
-import {createEntitesHandler} from '../helpers/entityRouteUtils';
+import {createEntitiesHandler} from '../helpers/entityRouteUtils';
 import express from 'express';
 import {isAuthenticatedForHandler} from '../helpers/auth';
 
 
 const router = express.Router();
-router.post('/create/handler', isAuthenticatedForHandler, createEntitesHandler);
+router.post('/create/handler', isAuthenticatedForHandler, createEntitiesHandler);
 export default router;
